agent/requirements.txt
paho-mqtt
taospy
redis